moved more code to data module
