+++ time_series/test_time_series/test_analysis.py
@@ -3,10 +3,9 @@ import pytest
 import autofit as af
 from time_series.analysis import Analysis
 from time_series.data import Data
-from time_series.fit import Fit
+from time_series.data import pdf
 from time_series.observable import Observable
 from time_series.species import Species, SpeciesObservables
-from time_series.util import pdf
 
 
 @pytest.fixture(name="a_0")
