+++ time_series/time_series/__init__.py
@@ -1,5 +1,5 @@
 from time_series.analysis import Analysis
-from time_series.data import Data
+from time_series.data import Data, generate_data, pdf
 from time_series.observable import Observable
 from time_series.observable import CompoundObservable
 from time_series.species import Species
