+++ time_series/time_series/data.py
@@ -1,4 +1,58 @@
+from random import randint
+
 import autofit as af
+import time_series as ts
+
+LOWER_LIMIT = 0
+UPPER_LIMIT = 20
+NUMBER_OF_POINTS = 400
+
+GRANULARITY = 100
+
+
+def rand_positive(upper_limit):
+    return randint(
+        0,
+        upper_limit * GRANULARITY
+    ) / GRANULARITY
+
+
+def generate_data(
+        number_of_observables,
+        number_of_species
+):
+    compound_observables = dict()
+    for number in range(number_of_observables):
+        compound_observables[
+            str(number)
+        ] = ts.CompoundObservable(
+            abundances=[
+                rand_positive(
+                    1
+                ) for _ in range(number_of_species)
+            ],
+            observables=[
+                ts.Observable(
+                    mean=rand_positive(3),
+                    deviation=rand_positive(2)
+                ) for _ in range(number_of_species)
+            ]
+        ).pdf(
+            LOWER_LIMIT,
+            UPPER_LIMIT,
+            NUMBER_OF_POINTS
+        )
+    return ts.Data(
+        **compound_observables
+    )
+
+
+def pdf(observable):
+    return observable.pdf(
+        LOWER_LIMIT,
+        UPPER_LIMIT,
+        NUMBER_OF_POINTS
+    )
 
 
 class Data(af.Dataset):
