+++ time_series/workspace/pipeline.py
@@ -1,55 +1,16 @@
-from random import randint
-
 import autofit as af
 import time_series as ts
 
 NUMBER_OF_SPECIES = 5
 NUMBER_OF_OBSERVABLES = 3
 
-LOWER_LIMIT = 0
-UPPER_LIMIT = 20
-NUMBER_OF_POINTS = 400
-
-GRANULARITY = 100
-
-
-def rand_positive(upper_limit):
-    return randint(
-        0,
-        upper_limit * GRANULARITY
-    ) / GRANULARITY
-
-
-def generate_data():
-    compound_observables = dict()
-    for number in range(NUMBER_OF_OBSERVABLES):
-        compound_observables[
-            str(number)
-        ] = ts.CompoundObservable(
-            abundances=[
-                rand_positive(
-                    1
-                ) for _ in range(NUMBER_OF_SPECIES)
-            ],
-            observables=[
-                ts.Observable(
-                    mean=rand_positive(3),
-                    deviation=rand_positive(2)
-                ) for _ in range(NUMBER_OF_SPECIES)
-            ]
-        ).pdf(
-            LOWER_LIMIT,
-            UPPER_LIMIT,
-            NUMBER_OF_POINTS
-        )
-    return ts.Data(
-        **compound_observables
-    )
-
 
 def run_phase():
     model = af.ModelMapper()
-    data = ts.Data()
+    data = ts.generate_data(
+        number_of_observables=NUMBER_OF_OBSERVABLES,
+        number_of_species=NUMBER_OF_SPECIES
+    )
 
     phase = af.Phase(
         analysis_class=ts.Analysis,
@@ -63,4 +24,4 @@ def run_phase():
 
 
 if __name__ == "__main__":
-    print(generate_data())
+    pass
